requirements.txt
PyOgg
PyOpenAL
PySoundFile
librosa